Rename next-prefixed folders
