--- openapi-spec-writer/mastra/blueprints/o6uye3483a515w5vqzlkdjil.json
@@ -1,47 +1,0 @@
-{
-  "title": "Make PR To Mastra",
-  "status": "PUBLISHED",
-  "createdAt": "2024-11-07T17:11:49.017Z",
-  "id": "o6uye3483a515w5vqzlkdjil",
-  "updatedAt": "2024-11-07T17:19:45.099Z",
-  "trigger": {
-    "id": "c3wfzadbjqupmj5lzz5e49ho",
-    "type": "PR_TO_MASTRA",
-    "payload": {}
-  },
-  "actions": [
-    {
-      "id": "s29dc3fpp3gdpovmwcm2dgdb",
-      "type": "ADD_TO_GIT",
-      "payload": {
-        "mastraConnectionId": "SYSTEM",
-        "yaml": "{{yaml_1730999911016}}",
-        "integration_name": "{{integration_name_1730999916310}}",
-        "site_url": "{{site_url_1730999926021}}",
-        "owner": "mastra-ai",
-        "repo": "mastra"
-      },
-      "variables": {
-        "yaml": {
-          "yaml_1730999911016": {
-            "refBlockId": "c3wfzadbjqupmj5lzz5e49ho",
-            "path": "yaml"
-          }
-        },
-        "integration_name": {
-          "integration_name_1730999916310": {
-            "refBlockId": "c3wfzadbjqupmj5lzz5e49ho",
-            "path": "integrationName"
-          }
-        },
-        "site_url": {
-          "site_url_1730999926021": {
-            "refBlockId": "c3wfzadbjqupmj5lzz5e49ho",
-            "path": "url"
-          }
-        }
-      },
-      "subActions": []
-    }
-  ]
-}
