--- openapi-spec-writer/mastra/blueprints/w4km3pdd6hpyvtmnd2z63bfx.json
@@ -1,59 +1,0 @@
-{
-  "title": "Open API Spec Gen",
-  "status": "PUBLISHED",
-  "createdAt": "2024-11-05T23:24:07.787Z",
-  "id": "w4km3pdd6hpyvtmnd2z63bfx",
-  "updatedAt": "2024-11-08T14:58:37.032Z",
-  "trigger": {
-    "id": "obbpq1y85urbhhhj7xdrx2c2",
-    "type": "WRITE_SPEC",
-    "payload": {}
-  },
-  "actions": [
-    {
-      "id": "ah6n1mlana6r9j2bni6wjq2f",
-      "type": "MINTLIFY_SITE_CRAWL",
-      "payload": {
-        "url": "{{url_1731014133434}}",
-        "mastraConnectionId": "SYSTEM",
-        "pathRegex": "{{pathRegex_1731077713204}}",
-        "limit": 3
-      },
-      "variables": {
-        "url": {
-          "url_1731014133434": {
-            "refBlockId": "obbpq1y85urbhhhj7xdrx2c2",
-            "path": "url"
-          }
-        },
-        "pathRegex": {
-          "pathRegex_1731077713204": {
-            "refBlockId": "obbpq1y85urbhhhj7xdrx2c2",
-            "path": "pathRegex"
-          }
-        }
-      },
-      "subActions": [
-        {
-          "id": "kgwyeluf5zbocv3gkcb1u34c",
-          "type": "GENERATE_MERGED_SPEC",
-          "parentActionId": "ah6n1mlana6r9j2bni6wjq2f",
-          "subActions": [],
-          "payload": {
-            "entityType": "CRAWL",
-            "mastra_entity_type": "{{mastra_entity_type_1731077515266}}",
-            "mastraConnectionId": "SYSTEM"
-          },
-          "variables": {
-            "mastra_entity_type": {
-              "mastra_entity_type_1731077515266": {
-                "refBlockId": "ah6n1mlana6r9j2bni6wjq2f",
-                "path": "entityType"
-              }
-            }
-          }
-        }
-      ]
-    }
-  ]
-}
