+++ openapi-spec-writer/mastra/index.ts
@@ -1,8 +1,8 @@
 import { Mastra } from "@mastra/core";
-import { agentOne } from "./next-agents";
-import { integrations } from "./next-integrations";
+import { agentOne } from "./agents";
+import { integrations } from "./integrations";
 // import * as syncs from './next-syncs';
-import * as tools from "./next-tools";
+import * as tools from "./tools";
 import { PostgresEngine } from "@mastra/engine";
 
 export const mastra = new Mastra<typeof integrations, typeof tools>({
